+++ apps/extension/entrypoints/keeper/keeper.ts
@@ -9,7 +9,7 @@ import {
   KeeperMessageTypes,
   type ZkProofResponse,
 } from "@evevault/shared";
-import { signWithIntent } from "@evevault/shared/wallet/signWithIntent";
+import { signWithIntent } from "@evevault/shared/wallet";
 import type { IntentScope } from "@mysten/sui/cryptography";
 import { Ed25519Keypair } from "@mysten/sui/keypairs/ed25519";
 import type { SuiChain } from "@mysten/wallet-standard";